Revert "Correct limit query return type to match spec ABI."
This reverts commit 7826d4ca2d3d620902b4c874a0f565457bdea0af.

Changing the parameter sizes breaks backward ABI.

+++ runtime/hsa-runtime/image/image_lut.h
@@ -65,13 +65,13 @@ class ImageLut {
 
   virtual Swizzle MapSwizzle(hsa_ext_image_channel_order32_t order) const = 0;
 
-  virtual size_t GetMaxWidth(hsa_ext_image_geometry_t geometry) const = 0;
+  virtual uint32_t GetMaxWidth(hsa_ext_image_geometry_t geometry) const = 0;
 
-  virtual size_t GetMaxHeight(hsa_ext_image_geometry_t geometry) const = 0;
+  virtual uint32_t GetMaxHeight(hsa_ext_image_geometry_t geometry) const = 0;
 
-  virtual size_t GetMaxDepth(hsa_ext_image_geometry_t geometry) const = 0;
+  virtual uint32_t GetMaxDepth(hsa_ext_image_geometry_t geometry) const = 0;
 
-  virtual size_t GetMaxArraySize(hsa_ext_image_geometry_t geometry) const = 0;
+  virtual uint32_t GetMaxArraySize(hsa_ext_image_geometry_t geometry) const = 0;
 
  private:
    DISALLOW_COPY_AND_ASSIGN(ImageLut);
